break out into more logical files + make some functions static
