+++ src/include/uibuilder.h
@@ -1,0 +1,7 @@
+#pragma once
+
+int lm_Realize(lua_State* L);
+
+int lm_Unrealize(lua_State* L);
+
+int lm_DestroyWidgetHierarchy(lua_State* L);
